test(bun-serve-static): split the /big stress loop into its own two files
The static-route stress loop pushes roughly 48GB over loopback (8 variants of
64 concurrent fetches x 24 iterations of a 4MB body). All of that shared the
180s per-file wall clock with the rest of bun-serve-static.test.ts, and on a
noisy AWS runner the /big section alone ran to ~195s in build 72253.

Move the stress loop into two files, one per res.body-access variant, each
getting its own 180s/360s budget. Same 4MB body, same 64x24 batch, same
toStrictEqual + RSS assertions, so the StaticRoute to_async/on_writable path
and the buffered-stream fast path stay fully covered.

Also in the shared helpers:
- drop the byteSize > 1MB branch (byteSize was always undefined, so the 48/10
  case has never been taken)
- fix randomAnyCaseLetter to write charCodeAt() instead of a string index,
  which had been filling /big with NUL bytes since #13704
- seed expected-durations.json so the LPT packer places the new files sensibly
